requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
PyPDF2
sentence-transformers
numpy
python-multipart
redis